+++ src/users_index/src/lib.rs
@@ -504,7 +504,7 @@ async fn request_cycles() -> Result<(), UserError> {
         ));
     }
 
-    transfer_cycles(CYCLES_TOP_UP_AMOUNT as u128, caller).await
+    transfer_cycles(CYCLES_TOP_UP_AMOUNT, caller).await
 }
 
 async fn transfer_cycles(cycles_amount: u128, caller: Principal) -> Result<(), UserError> {
